Day 3 solution made a little nicer

+++ 3/puzzle1b.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#define BITS (8 * sizeof(unsigned))
+#define BITS (8 * ((int)sizeof(unsigned)))
 
 int parse(FILE *stream, unsigned **numbersp, size_t *sizep)
 {
